+++ src/plugins/stake-plugins/stakePlugins.ts
@@ -3,7 +3,7 @@ import { makeTronStakePlugin } from './currency/tronStakePlugin'
 import { makeGenericStakePlugin } from './generic/GenericStakePlugin'
 import { genericPlugins } from './generic/pluginInfo'
 import { makeTcSaversPlugin } from './thorchainSavers/tcSaversPlugin'
-import { makeThorchainTcyPlugin } from './thorchainSavers/tcyPlugin'
+import { makeTcSaversPluginSegwit } from './thorchainSavers/tcSaversPluginSegwit'
 import { StakePlugin } from './types'
 import { makeUniV2StakePlugin } from './uniswapV2/uniV2Plugin'
 
@@ -23,7 +23,7 @@ export const getStakePlugins = async (pluginId: string): Promise<StakePlugin[]> 
     makeTcSaversPlugin(pluginId, { initOptions: tcInitOptions }).catch(e => {
       console.warn(e.message)
     }),
-    makeThorchainTcyPlugin(pluginId, { initOptions: tcInitOptions }).catch(e => {
+    makeTcSaversPluginSegwit(pluginId, { initOptions: tcInitOptions }).catch(e => {
       console.warn(e.message)
     }),
     makeTronStakePlugin(pluginId).catch(e => {